requirements.txt
flask
python-dotenv
requests
python-telegram-bot==20.0
